Finished carbon value tests
2 of 61 tests are failing, one by .01, and one because of an issue probably due to initial conditions in RAS modules.
